requirements.txt
fastapi
uvicorn[standard]
python-dotenv

numpy==1.26.4
pandas==2.1.4
scikit-learn==1.3.2
scipy==1.11.4

faiss-cpu==1.7.4

torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2

sentence-transformers==2.2.2
transformers==4.35.2

requests
